Implemented a basic syntax according to the syntax protocol specification (which I haven't written yet). The current implementation should be improved upon, but it basically shows how to do it.
Also implemented a demo command that accepts a string and inserts its
reverse in the buffer.  This shows that the words in the buffer are
actually presentations (of type string) that become clickable by the
accept.

Added two missing methods on region-to-sequence.  There were no
methods when one of the arguments is an offset instead of a mark.