Factored aspects of named objects (currently buffers and syntaxes) into a syntax-mixin class in base.lisp.  Updated packages.lisp accordingly.
Implemented syntax completion (i.e., the possibility to use CLIM
completion to determine the name of a syntax).  I

Implemented an extended command "Set Syntax" using the completion.
Currently, it does not invalidate the CLIM output history, because I
need to think a bit more about how to do that properly.