Alternative syntax for editing common lisp (called "Lisp").
I have not completely lost hope of making the Earley-based syntax
module fast enough, but as of now I don't know how.

This syntax module provides an incremental resynchronizing LR parser
for Common Lisp.  It is fast enough the parse tree for the entire
buffer is kept updated, as opposed to just for what is visible in the
CLIM pane.

Ultimately, I hope to factor out the parts that are not specific to
Lisp into a different module, perhaps the existing syntax.lisp file.
I'll be interested in your opinions.