Improvements to Lisp syntax. (thanks to John Q Splittist)
Here is his own description of these improvements:

This patch:

* fixes presentations of multi-token symbols and strings
* introduces a new presentation type, the 'unknown-symbol, for symbol
tokens that haven't got a package in the image (because, eg. the file
hasn't been loaded)
* introduces a new presentation type, the 'lisp-string, for strings in
the file surrounded by #\"s
* presents every token as a 'string.

Also included is a presentation translator from 'lisp-string to 'string
that doesn't work. It ought to, and I seem to have got back into the
gesture/pointer-event code with things still making (to me) sense, so
I'd be grateful if someone could check whether it works for them.

Things to play with:

* M-x Accept String (most things mouseable)
* M-x Accept Symbol (see what the system can find, and where - 'symbols
are returned as the actual symbol; 'unknown-symbols are returned as strings
* M-x Accept Lisp String (source code strings are mouseable)
* M-% [being Query Replace], then mouse and click to choose the strings!

Things to think about:

* Should 'string be for actual lisp strings, and (say) ESA-string (or
editor-string) be for sequences of objects in the buffer? This makes
sense to me, as some commands that accept a sequence of objects from the
buffer might be usable in non-text-editor contexts. (Simply changing
commands like com-query-replace from (accept 'string ...) to (accept
'esa-string ...), and changing a couple of things in lisp-syntax, would
work.)
* What other things might it be useful to mouse around with?
* Is there a natural meaning for simply clicking on something in the buffer?

Things to do:

* (still!) Numbers
* work out why the presentation translator isn't working...